+++ anyblok_pyramid/tests/test_argsparse.py
@@ -1,4 +1,6 @@
-from anyblok_pyramid._argsparse import define_wsgi_option
+from anyblok_pyramid._argsparse import (define_wsgi_option,
+                                        define_beaker_option,
+                                        define_wsgi_debug_option)
 from anyblok.tests.testcase import TestCase
 from anyblok.tests.test_argsparse import MockArgumentParser
 
@@ -13,7 +15,16 @@ class TestArgsParseOption(TestCase):
         cls.configuration = {}
         cls.function = {
             'define_wsgi_option': define_wsgi_option,
+            'define_beaker_option': define_beaker_option,
+            'define_wsgi_debug_option': define_wsgi_debug_option,
         }
 
     def test_define_wsgi_option(self):
         self.function['define_wsgi_option'](self.parser, self.configuration)
+
+    def test_define_beaker_option(self):
+        self.function['define_beaker_option'](self.parser, self.configuration)
+
+    def test_define_wsgi_debug_option(self):
+        self.function['define_wsgi_debug_option'](self.parser,
+                                                  self.configuration)